[chore] 將 teacher infos API 的使用從後台個人資訊頁面搬到後台主頁
